Decouple some of the XML code into a separate file
